feat: Initial session support.
Right now the session doesn't track QoS or subscriptions, but already enables making sure that only
one client with a given name can connect at a time. It took a while to come up with a good race-free
architecture, and this one seems satisfying as it makes sure that only one SessionData instance
exists for a given client.name, and that instance gets moved from the live client to the session
store as needed.