feat: Publish using an increasing Pid instead of a static one.
The prev_pid is stored in the session, and pid assignment makes sure to avoid currently-used
pids. We now send a Msg::publish instead of a Msg::PkgOut to the subscriber client, so that encoding
is done by the subscriber, because that's where the correct session.prev_pid is.

Forgot to commit session.rs in previous commits, since 800eb21cc8b7597a8fa72f303b59d28d10112385 two
weeks ago. Woops.