doc: Add LICENSE.txt and CONTRIBUTING.md, tweak README.md and Cargo.toml
This is in preparation for an opensource release :) The readme got a big update of current status
and plans, the toml got metadata for crates.io. The license was chosen to be permissive and in line
with the main deps. As requested, added a mention for my employer who sponsored the work.