feat: Initial support for external decoder command.
That command can be used to convert from a binary format, amend a json format, or whatever
transformation you may need.

This is an MVP of the feature: it currently blocks the client thread indefinitely, doesn't support
passing arguments to the command, and starts a new command each time (making stateful operation
harder). I spent a good bit of time trying to do a proper async implementation, but it looks like
it'll be easyer if I upgrade tokio first.