chore: Release 0.1.0
First release :)

Mqttest is at this stage a niffty unittesting tool, even if some features are still missing. My
short-term plan is to work on the library usecase (as opposed to running a binary). Now is a great
time to chime in with your usecase, bug report, or code contribution.