refactor/doc: Make the libray pub API more usable, improve docs, add  a unittest.
The start() and run() methods have been merged into a single step, and we now expose a JoinHandle to
wait for the server to finish and retrieve the result. There is now a unittest to make sure that
this API makes sense and is convenient. The docs have seen a big update.