deps: Update tokio and related crates
Long overdue, to catch up with API/ecosystem changes (thankfully simple enough at our scale) and
security lints. This brings an MSRV bump to 1.65, although we'll probably raise it a bit more later
for convenience.

We switched to a specific mqttrs commit that has a needed `bytes` update, but not to the newest
mqttrs crate yet (too many changes required). Might be simpler to propose a 0.5 that re-adds an API
with owned structs.